style(api): unify procedure schemas in a single file for simplified middleware configuration
