+++ packages/api/src/middlewares/withPagination.ts
@@ -4,16 +4,16 @@ import { t } from "../trpc-client";
 
 export const DEFAULT_PAGE_LIMIT = 25;
 
-export const paginationSchema = z.object({
+export const withPaginationSchema = z.object({
   p: z.number().optional(),
   ps: z.number().optional(),
 });
 
-export type PaginationSchema = z.infer<typeof paginationSchema>;
+export type WithPaginationSchema = z.infer<typeof withPaginationSchema>;
 
 export const withPagination = t.middleware(({ next, input }) => {
   const { p: offset = 1, ps: limit = DEFAULT_PAGE_LIMIT } =
-    paginationSchema.optional().parse(input) ?? {};
+    withPaginationSchema.optional().parse(input) ?? {};
 
   return next({
     ctx: {
