--- packages/api/src/routers/blockchain-sync-state/getState.schema.ts
@@ -1,10 +1,0 @@
-import { z } from "@blobscan/zod";
-
-export const getStateInputSchema = z.void();
-
-export const getStateOutputSchema = z.object({
-  lastAggregatedBlock: z.number().nullable(),
-  lastFinalizedBlock: z.number().nullable(),
-  lastLowerSyncedSlot: z.number().nullable(),
-  lastUpperSyncedSlot: z.number().nullable(),
-});
