+++ packages/api/src/utils/schemas.ts
@@ -14,8 +14,6 @@ const zodRollupEnums = [
   "zora",
 ] as const;
 
-const zodExpandEnums = ["blob", "blob_data", "block", "transaction"] as const;
-
 /**
  * This is a type-safe way to get the enum values as we can't use `Object.values`
  * on zod enums directly
@@ -51,8 +49,6 @@ if (missingBlobStorageEnums.length) {
 
 export type ZodRollupEnum = (typeof zodRollupEnums)[number];
 
-export type ZodExpandEnum = (typeof zodExpandEnums)[number];
-
 export type ZodBlobStorageEnum = (typeof zodBlobStorageEnums)[number];
 
 export const blobStorageSchema = z.enum(zodBlobStorageEnums);
@@ -73,19 +69,3 @@ export const blockNumberSchema = z.number().nonnegative();
 export const slotSchema = z.number().nonnegative();
 
 export const blobIndexSchema = z.number().nonnegative();
-
-export const expandSchema = z
-  .string()
-  .refine(
-    (value) => {
-      const values = value.split(",");
-
-      return values.every((v) => zodExpandEnums.includes(v as ZodExpandEnum));
-    },
-    {
-      message: `Invalid 'expand' value. It must be a comma separated list of the following values: ${zodExpandEnums.join(
-        ", "
-      )}`,
-    }
-  )
-  .transform((value) => value.split(",") as ZodExpandEnum[]);
